GetPlayerInfo( ) Fixes. No extra \n
GetPlayerInfo( ) now no longer tries to add an int to a string with out
first properly converting it.
GetPlayerInfo( ) now will only try to append if the tempPlayer is !=
NULL.
Removed the extra \n at the end of GetPlayerInfo( ).
Added some logic so that there will be no \n at the end of the strings
returned by SearchByClass( ) and SearchByHats( )